refactor:♻️CacheStoreインターフェースをinfrastructure/cacheに移動(#4)
