+++ internal/infrastructure/cache/cache_store.go
@@ -1,5 +1,5 @@
-// Package port はアプリケーション層の外部依存インターフェースを定義する
-package port
+// Package cache はキャッシュストアの操作を定義する
+package cache
 
 import (
 	"context"
@@ -11,7 +11,7 @@ type CacheStore interface {
 	// Get はキーから値を取得する
 	Get(ctx context.Context, key string) (string, error)
 
-	// Set はキーに値を設定する（TTL付き）
+	// Set はキーに値を設定する(TTL付き)
 	Set(ctx context.Context, key string, value string, expiration time.Duration) error
 
 	// Delete はキーを削除する
